Punten systeem koppelen aan rewards
Punten systeem koppelen aan rewards + enkele aanpassingen aan Gebruiker en scorebord.

+++ src/main/java/hhs/proj2_klas6_groep6d/Scorebord.java
@@ -23,7 +23,7 @@ public class Scorebord {
     }
 
     public ArrayList<Gebruiker> orderByPoints(){
-        Comparator<Gebruiker> comparator = Comparator.comparingInt(Gebruiker::getPuntenSaldo);
+        Comparator<Gebruiker> comparator = Comparator.comparingDouble(Gebruiker::getPuntenSaldo);
         scorebord.sort(comparator);
         return scorebord;
     }
